refactor: extract inline tests from 8 files to reduce module sizes (#437)
Move #[cfg(test)] mod tests blocks from 8 source files into separate
_tests.rs files, following the existing pattern used by web_fetch.rs
and web_search.rs. This reduces production code LOC without changing
any test behavior.

Files processed:
- crates/tools/src/bash_output.rs (650 -> 142 LOC)
- crates/tools/src/todo_write.rs (632 -> 183 LOC)
- crates/tools/src/kill_shell.rs (492 -> 108 LOC)
- crates/tools/src/agent_registry.rs (538 -> 289 LOC)
- crates/cli/src/plugins/agent_discovery.rs (827 -> 265 LOC)
- crates/cli/src/plugins/frontmatter_substitution.rs (721 -> 323 LOC)
- crates/cli/src/update/mod.rs (612 -> 49 LOC)
- crates/cli/src/plugins/manifest.rs (585 -> 300 LOC)

Also updates zero_bs_compliance.rs to exclude _tests.rs files from
production code quality checks.
